Synchronous reset
While all other register in the FPGA implementation have synchronous reset, this one had an asynchronous reset. This could add an unnecessary asynchronous input timing constraint for synthesis. It turns out this reset is not required at all, as `word_cnt_s` is set to zero in state `state_s = IDLE` and does not generate any of the flow control outputs in that state.